Tidied some docs, made VC++ 6 DLL compilation work with wxvc_dll.dsp, removed wxvc6.dsp, updated projgen, small fixes for OGL
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4286 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ utils/ogl/samples/studio/studio.cpp
@@ -492,10 +492,10 @@ bool csApp::WriteOptions()
 {
     wxConfig config("OGL Studio", "wxWindows");
 
-    config.Write("mainX", m_mainFramePos.x);
-    config.Write("mainY", m_mainFramePos.y);
-    config.Write("mainWidth", m_mainFrameSize.x);
-    config.Write("mainHeight", m_mainFrameSize.y);
+    config.Write("mainX", (long) m_mainFramePos.x);
+    config.Write("mainY", (long) m_mainFramePos.y);
+    config.Write("mainWidth", (long) m_mainFrameSize.x);
+    config.Write("mainHeight", (long) m_mainFrameSize.y);
     config.Write("gridStyle", (long) m_gridStyle);
     config.Write("gridSpacing", (long) m_gridSpacing);
